app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
psycopg2-binary==2.9.10
databricks-sdk==0.40.0
markdown==3.7
python-multipart==0.0.20
httpx==0.28.1
